Time: 0 ms (100.00%), Space: 7 MB (61.35%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.cpp
@@ -1,18 +1,17 @@
 class Solution {
 public:
     typedef long long int ll;
-    bool isPerfectSquare(long double x){
+    bool isPerfectSquare(double x){
         if (x >= 0) {
-            long long sr = sqrt(x);
+            int sr = sqrt(x);
             return (sr * sr == x);
         }
         return false;
     }
     bool judgeSquareSum(int c) {
         for(ll i=0;i*i<=c;i++){
-            long double a=c-(i*i);
+            double a=c-(i*i);
             if(isPerfectSquare(a)){
-                //cout<<a<<endl;
                 return true;
             }
         }
